closes #81: Remove hooks in both supervisor and systick ISRs

+++ projects/lpc40xx_freertos/l3_drivers/pwm1.h
@@ -16,4 +16,4 @@ typedef enum {
 
 /// For simple single edge PWM, all channels share the same frequency
 void pwm1__init_single_edge(uint32_t frequency_in_hertz);
-void pwm1__set_duty_cycle(pwm1_channel_e pwm1_channel, double duty_cycle_in_percent);
+void pwm1__set_duty_cycle(pwm1_channel_e pwm1_channel, float duty_cycle_in_percent);
